% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm_combo_cov_data.R
\docType{data}
\name{blrm_combo_cov_data}
\alias{blrm_combo_cov_data}
\title{A hypothetical dual-agent trial data with covariates}
\format{
A patient-wise trial data:
\describe{
  \item{Scenario}{Hypothetical scenarios}
  \item{DSPX1}{dose of drug1 tested for every patient}
  \item{DSPX2}{dose of drug2 tested for every patient}
  \item{DLT}{If DLT was observed for every patient}
  \item{Azole}{A dummy variable}
  ...
}
}
\usage{
data(blrm_combo_cov_data)
}
\description{
A simple \code{data.frame} that is used for testing and running of examples
}
\examples{
data(blrm_combo_cov_data)
blrm_combo_cov_data
}
\keyword{datasets}
