% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm_combo_data.R
\docType{data}
\name{blrm_combo_data}
\alias{blrm_combo_data}
\title{A hypothetical Dual-agent trial data}
\format{
A cohort-wise trial data:
\describe{
  \item{Scenario}{Hypothetical scenarios}
  \item{DSPX1}{doses of drug1 tested for every cohort, separated by comma}
  \item{DSPX2}{doses of drug2 tested for every cohort, separated by comma}
  \item{npat}{the number of patients treated for every cohort, separated by comma}
  \item{nDLT}{the number of DLTs observed for every cohort, separated by comma}
  ...
}
}
\usage{
data(blrm_combo_data)
}
\description{
A simple \code{data.frame} that is used for testing and running of examples
}
\examples{
data(blrm_combo_data)
blrm_combo_data
}
\keyword{datasets}
