% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm_mono_cov_data.R
\docType{data}
\name{blrm_mono_cov_data}
\alias{blrm_mono_cov_data}
\title{A hypothetical single-agent trial data with covariates}
\format{
A patient-wise trial data:
\describe{
  \item{Scenario}{Hypothetical scenarios}
  \item{DSPX}{dose tested for every patient}
  \item{DLT}{If DLT was observed for every patient}
  \item{Azole}{A dummy variable}
  ...
}
}
\usage{
blrm_mono_cov_data
}
\description{
A simple \code{data.frame} that is used for testing and running of examples
}
\examples{
blrm_mono_cov_data
}
\keyword{datasets}
