% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm_mono_data.R
\docType{data}
\name{blrm_mono_data}
\alias{blrm_mono_data}
\title{A hypothetical single-agent trial data}
\format{
A cohort-wise trial data:
\describe{
  \item{Scenario}{Hypothetical scenarios}
  \item{DSPX}{doses tested for every cohort, separated by comma}
  \item{npat}{the number of patients treated for every cohort, separated by comma}
  \item{nDLT}{the number of DLTs observed for every cohort, separated by comma}
  ...
}
}
\usage{
blrm_mono_data
}
\description{
A simple \code{data.frame} that is used for testing and running of examples
}
\examples{
blrm_mono_data
}
\keyword{datasets}
