% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodel-name.R
\docType{package}
\name{demodel}
\alias{demodel}
\title{demodel: A package for early phase dose finding trial}
\description{
The demodel package provides four categories of important functions:
demodelFit, MBInfo, BayesInfo and demodelControl.
}
\section{demodel functions}{

The demodel functions ...
}

