% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodelControl.R
\name{demodelControl}
\alias{demodelControl}
\title{Input parameters about output report}
\usage{
demodelControl(
  core = parallel::detectCores(),
  date = format(Sys.Date(), "\%Y-\%m-\%d"),
  code.name = NULL,
  fig.height = 8,
  fig.width = 8,
  output.path = getwd(),
  table.file.name = NULL,
  fig.file.name = NULL
)
}
\arguments{
\item{core}{A integer indicating how many cores will be used for parallel computation. The default uses all available cores.}

\item{date}{The default is the the date when one run this code}

\item{code.name}{the R file name used for \code{\link{demodelFit}} function. The default is 'demodel.R'. Can be modified for version control.}

\item{fig.height, fig.width}{A integer. The height and width of figures attached in the final report. The default is 8 for both width and height. See \code{\link[ggplot2]{ggsave}} for more information.}

\item{output.path}{A string. Folder location where the report is saved}

\item{table.file.name}{A string with suffix .xlsx. The file name of final report.}

\item{fig.file.name}{deprecated. Will be deleted later.}
}
\value{
A named list
}
\description{
Input parameters about output report
}
\examples{

\dontrun{

demodelControl(core = 8,
               date = format(Sys.Date(), "\%Y-\%m-\%d"),
               code.name = "demodel_core.R",
               output.path = getwd(),
               table.file.name = "preview.xlsx",
               fig.file.name = NULL)


}
}
