% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_check.R
\name{formula_check}
\alias{formula_check}
\title{create formula used in Bayesian modeling}
\usage{
formula_check(formula)
}
\arguments{
\item{formula}{A string. See example.}
}
\value{
A named list with 4 elements
}
\description{
create formula used in Bayesian modeling
}
\examples{

\dontrun{

formula.check(paste(DLT ~ DSPX | Azole))

}
}
