% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBInfo.R
\name{MBInfo}
\alias{MBInfo}
\title{Input information about model-based design}
\usage{
MBInfo(
  dose.levels = NULL,
  ref.dose = NULL,
  bounds = c(0.16, 0.33),
  ewoc = 0.25,
  trial.name = NULL,
  drug.name = NULL,
  drug.unit = NULL
)
}
\arguments{
\item{dose.levels}{A named list. names must match drug names.}

\item{ref.dose}{A vector of numeric value.}

\item{bounds}{A vector of two numeric values. Cutoff points of two categories. By the default, (0.16, 0.33).}

\item{ewoc}{A numeric value. Escalation with overdosing control. The default is 0.25.}

\item{trial.name}{A string.}

\item{drug.name}{A vector of string.}

\item{drug.unit}{A vector of string.}
}
\value{
A named list
}
\description{
Input information about model-based design
}
\examples{

\dontrun{
  MB.Info <- MBInfo(dose.levels = list(Drug = c(80, 120, 160, 200, 240, 280)),
                    ref.dose = 120,
                    bounds = c(0.16, 0.33),
                    ewoc = 0.25,
                    trial.name = "Drug666",
                    drug.name = "DrugX",
                    drug.unit = "mg QD")
}
}
