% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodel-name.R
\docType{package}
\name{demodel}
\alias{demodel}
\title{demodel: A package for computating the notorious bar statistic}
\description{
The demodel package provides two categories of important functions:
demodelFit and demodelControl.
}
\section{demodel functions}{

The demodel functions ...
}

